requirements.txt
python-telegram-bot
requests
python-dotenv
SQLite3-0611
openai
pymongo